requirements.txt
bs4
Flask
Flask-SocketIO
requests
python-dotenv